Fix broken links in Ethernet.md
ICS chip pdf was a dead link and for some reason BCM datasheet link was to the ICS datasheet. Not sure if this is significant but the link I provided is to the IDT website and the ICS chip pdf has IDT's document headers instead of ICS, and the final "Ordering Information" page has been changed, and the IDT PDF has working links, other than these I do not believe there are any differences in the documents.